pyproject.toml
[project]
name = "porcurment"
version = "0.1.0"
description = "AI-powered procurement deal finder using CrewAI agents"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.40.0",
    "pandas>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
